Restructured dashboard/controller/dashboard
